+++ js/package.json
@@ -11,7 +11,7 @@
   },
   "dependencies": {
     "@flarum/prettier-config": "^1.0.0",
-    "flarum-tsconfig": "^1.0.2",
+    "flarum-tsconfig": "^2.0.0",
     "flarum-webpack-config": "^3.0.0",
     "fuzzyset": "^1.0.7",
     "lodash.debounce": "^4.0.8",
