+++ js/src/forum/components/ReactionsModal.tsx
@@ -110,7 +110,10 @@ export default class ReactionsModal extends Modal<ReactionsModalAttrs> {
   async load(): Promise<void> {
     this.loading = true;
 
-    const response = await app.store.find<ApiResponsePlural<PostReaction>>(`/posts_reactions`, { include: 'user,reaction', filter: { post: this.attrs.post.id() } });
+    const response = await app.store.find<ApiResponsePlural<PostReaction>>(`/posts_reactions`, {
+      include: 'user,reaction',
+      filter: { post: this.attrs.post.id() },
+    });
 
     const groupedReactions = groupBy(response, (r: PostReaction) => r.reactionId());
     const reactions: ReactionGroup[] = [];
